+++ ios/AppLovinMAX.m
@@ -655,6 +655,12 @@ RCT_EXPORT_METHOD(destroyMRec:(NSString *)adUnitIdentifier)
 RCT_EXPORT_METHOD(loadInterstitial:(NSString *)adUnitIdentifier)
 {
     MAInterstitialAd *interstitial = [self retrieveInterstitialForAdUnitIdentifier: adUnitIdentifier];
+    if ( !interstitial )
+    {
+        [self sendReactNativeEventForAdLoadFailed: @"OnInterstitialLoadFailedEvent" forAdUnitIdentifier: adUnitIdentifier withError: nil];
+        return;
+    }
+    
     [interstitial loadAd];
 }
 
@@ -681,6 +687,12 @@ RCT_EXPORT_METHOD(setInterstitialExtraParameter:(NSString *)adUnitIdentifier :(N
 RCT_EXPORT_METHOD(loadRewardedAd:(NSString *)adUnitIdentifier)
 {
     MARewardedAd *rewardedAd = [self retrieveRewardedAdForAdUnitIdentifier: adUnitIdentifier];
+    if ( !rewardedAd )
+    {
+        [self sendReactNativeEventForAdLoadFailed: @"OnRewardedAdLoadFailedEvent" forAdUnitIdentifier: adUnitIdentifier withError: nil];
+        return;
+    }
+    
     [rewardedAd loadAd];
 }
 
@@ -745,14 +757,23 @@ RCT_EXPORT_METHOD(setRewardedAdExtraParameter:(NSString *)adUnitIdentifier :(NSS
     [self sendReactNativeEventWithName: name body: [self adInfoForAd: ad]];
 }
 
-- (void)handleNativeAdLoadFailureForAdUnitIdentifier:(NSString *)adUnitIdentifier error:(nullable MAError *)error
+- (void)sendReactNativeEventForAdLoadFailed:(NSString *)name forAdUnitIdentifier:(NSString *)adUnitIdentifier withError:(nullable MAError *)error
 {
-    [self sendReactNativeEventWithName: @"OnNativeAdLoadFailedEvent"
-                                  body: @{@"adUnitId" : adUnitIdentifier,
-                                          @"code" : @(error.code),
+    NSMutableDictionary *body = @{@"adUnitId": adUnitIdentifier}.mutableCopy;
+    
+    if ( error )
+    {
+        [body addEntriesFromDictionary: @{@"code" : @(error.code),
                                           @"message" : error.message,
                                           @"adLoadFailureInfo" : error.adLoadFailureInfo ?: @"",
                                           @"waterfall": [self createAdWaterfallInfo: error.waterfall]}];
+    }
+    else
+    {
+        [body addEntriesFromDictionary: @{@"code" : @(MAErrorCodeUnspecified)}];
+    }
+    
+    [self sendReactNativeEventWithName: name body: body];
 }
 
 - (void)didFailToLoadAdForAdUnitIdentifier:(NSString *)adUnitIdentifier withError:(MAError *)error
@@ -782,11 +803,7 @@ RCT_EXPORT_METHOD(setRewardedAdExtraParameter:(NSString *)adUnitIdentifier :(NSS
         return;
     }
     
-    [self sendReactNativeEventWithName: name body: @{@"adUnitId" : adUnitIdentifier,
-                                                     @"code" : @(error.code),
-                                                     @"message" : error.message,
-                                                     @"adLoadFailureInfo" : error.adLoadFailureInfo ?: @"",
-                                                     @"waterfall": [self createAdWaterfallInfo: error.waterfall]}];
+    [self sendReactNativeEventForAdLoadFailed: name forAdUnitIdentifier: adUnitIdentifier withError: error];
 }
 
 - (void)didClickAd:(MAAd *)ad